Refactor user and product models and move them to the appropriate directories, changed working dir
