+++ backend/app/Dockerfile.dev
@@ -6,19 +6,20 @@ RUN apt-get update && \
     apt-get install -y nano --no-install-recommends && \
     rm -rf /var/lib/apt/lists/*
 # Set the working directory in the container
-WORKDIR /code/app
+WORKDIR /code
 
+RUN mkdir app
 # Copy the requirements file to the working directory
-COPY requirements.txt .
+COPY requirements.txt ./app
 
 # Install the dependencies
-RUN pip install --no-cache-dir -r requirements.txt
+RUN pip install --no-cache-dir -r ./app/requirements.txt
 
-# Copy the application code to the working directory
-COPY . .
+# # Copy the application code to the working directory
+COPY . ./app
 
 # # Expose the port FastAPI will run on
 EXPOSE 8000
 
 # Command to run the FastAPI server
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
